requirements.txt
python-whois==0.6.3
requests==2.11.1
